% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{arcs_to_seminr}
\alias{arcs_to_seminr}
\title{Convert bnlearn Arcs to seminr Paths}
\usage{
arcs_to_seminr(learned_graph)
}
\arguments{
\item{learned_graph}{A bnlearn object (bn).}
}
\value{
A seminr structural model object.
}
\description{
Convert bnlearn Arcs to seminr Paths
}
\keyword{internal}
