% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{cIPMA}
\alias{cIPMA-package}
\alias{cIPMA}
\title{cIPMA: Combined Importance-Performance Map Analysis}
\description{
The cIPMA package allows researchers to conduct a combined Importance-Performance Map Analysis
by integrating results from PLS-SEM (using seminr) and Necessary Condition Analysis (using NCA).
}
\references{
Hauff, S., Richter, N. F., Sarstedt, M., & Ringle, C. M. (2024).
Importance and performance in PLS-SEM and NCA: Introducing the combined importance-performance map analysis (cIPMA).
Journal of Retailing and Consumer Services, 78, 103723.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarekDejaUJ/cIPMA}
  \item Report bugs at \url{https://github.com/MarekDejaUJ/cIPMA/issue}
}

}
