% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipma_analysis.R
\name{print.cipma_assumptions}
\alias{print.cipma_assumptions}
\title{Print Assumption Check Results}
\usage{
\method{print}{cipma_assumptions}(x, ...)
}
\arguments{
\item{x}{A cipma_assumptions object.}

\item{...}{Ignored.}
}
\description{
Print Assumption Check Results
}
