% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{rescale_0_100}
\alias{rescale_0_100}
\title{Rescale Data to 0-100 Range}
\usage{
rescale_0_100(x, min_val, max_val)
}
\arguments{
\item{x}{Numeric vector.}

\item{min_val}{Theoretical minimum of the scale.}

\item{max_val}{Theoretical maximum of the scale.}
}
\value{
A numeric vector rescaled between 0 and 100.
}
\description{
This helper function rescales vector data based on theoretical minimum and maximums.
}
\keyword{internal}
